Refactor home widget
